Clear training process
