+++ recognition/vision_transformer_47147571/dataset.py
@@ -80,3 +80,4 @@ class ADNIDataset(Dataset):
         
         label = self.labels[idx]
         return image, label
+
